+++ edje-wp.php
@@ -5,7 +5,7 @@ Description: Collection of code to help developers customize WordPress into full
 Plugin URI: http://github.com/hrsetyono/edje-wp
 Author: The Syne Studio
 Author URI: http://thesyne.com/
-Version: 0.5.3
+Version: 0.5.3b
 */
 
 // exit if accessed directly
@@ -13,11 +13,10 @@ if(!defined('ABSPATH') ) { exit; }
 
 // Constant
 define('H_URL', plugin_dir_url(__FILE__) );
-define('H_DIR', basename(dirname(__FILE__) ).'/'.basename(__FILE__) );
+define('H_BASE', basename(dirname(__FILE__) ).'/'.basename(__FILE__) );
 
 // enable checking if plugin active
 include_once(ABSPATH . 'wp-admin/includes/plugin.php');
-
 
 require_once 'vendor/all.php';
 require_once 'lib/all.php';