+++ lib/h-on-install.php
@@ -4,7 +4,7 @@ new H_Install();
 class H_Install {
 
   function __construct() {
-    register_activation_hook(H_DIR, array($this, 'activation_hook') );
+    register_activation_hook(H_BASE, array($this, 'activation_hook') );
   }
 
   function activation_hook() {